feat(dsl): 画面値の読み出し・Sel の対称化・秒の小数化とスレッド違反の非致命化
Swift DSL の評価で挙がった4点に対応する。

- `exist` の戻り値から `.text`/`.value`/`.id` を読めるようにする。新コマンドは足さず、
  `exist` が既に解決した要素を持ち帰るだけなので追加のデバイス往復もステップ記録も無い。
  値は成功時のみ載せる(掴めなかったのに値が読める状態を作らない)
- セレクタを取るコマンドの String / Sel を1対1にする。Shirates 由来の別名族
  (tapWithScrollDown 等8個)に Sel 版が無く、型付き経路を選ぶと機能が減る状態だった
- 秒引数を全て Double にする(`timeout: 1.2` が書ける)。表示は FTSeconds.format に集約し、
  ifCanSelect のポーリングも 0.5 秒固定をやめてサブ秒が丸められないようにする
- DSL スレッド外からの呼び出しで fatalError せず、失敗ステップを1回だけ記録して中断する。
  1プロセス=1シナリオなのでここで落とすとレポートごと消えていた。
  対になる排他(record/中断フラグ・FTRuntime の core 参照)を入れ、正しさは TSan で確認した

実行プロファイルの defaultTimeout が小数を取り得るようになるため、旧拡張が正当な値を
弾かないようプロトコル版を +1 する。

+++ Sources/FTCore/RunProfile.swift
@@ -214,8 +214,8 @@ public struct RunProfileDocument: Codable, Sendable, Equatable {
     public var screenIs: Bool?
     /// レポート出力先(プロジェクトルート相対 or 絶対。既定 "reports")
     public var reportDir: String?
-    /// DSL コマンドの既定タイムアウト秒(省略時は DSL 側の既定値)
-    public var defaultTimeout: Int?
+    /// DSL コマンドの既定タイムアウト秒(小数可。省略時は DSL 側の既定値)
+    public var defaultTimeout: Double?
     /// シナリオ単位の壁時計タイムアウト秒(ホスト側 watchdog。子には渡さない。省略時 90)。
     /// defaultTimeout(子内部の検証待ち)とは別物
     public var scenarioTimeout: Int?
@@ -268,7 +268,7 @@ public struct RunProfileDocument: Codable, Sendable, Equatable {
 
     public init(app: String? = nil, devices: [RunDeviceRef]? = nil, fm: Bool? = nil,
                 heal: Bool? = nil, falsePositiveCheck: Bool? = nil, screenIs: Bool? = nil,
-                reportDir: String? = nil, defaultTimeout: Int? = nil, scenarioTimeout: Int? = nil,
+                reportDir: String? = nil, defaultTimeout: Double? = nil, scenarioTimeout: Int? = nil,
                 machine: String? = nil, iosInappEngine: Bool? = nil,
                 wipeDataOnBloat: Bool? = nil, wipeDataThresholdGB: Double? = nil,
                 recoverCpuFallbackToGpu: Bool? = nil,
@@ -356,7 +356,7 @@ public struct ResolvedProfile: Sendable {
     public var heal: Bool { fm.heal }
     /// 絶対パス解決済み
     public let reportDir: URL
-    public let defaultTimeout: Int?
+    public let defaultTimeout: Double?
     /// シナリオ単位の壁時計タイムアウト秒(ホスト側 watchdog。nil=未指定→run 側で既定 90 を適用)
     public let scenarioTimeout: Int?
     /// 実行開始時に Android AVD 肥大化を Wipe Data するか(既定 true)
